Crate gstreamer_sdp

Modules

prelude

Structs

Continue
Continue calling the closure in the future iterations or drop it.

Error
A generic error capable of representing various error domains (types).

TypedValue
A statically typed Value.

Value
A generic value capable of carrying various types.

Enums

MIKEYCacheType
The different cache types

MIKEYEncAlg
The encryption algorithm used to encrypt the Encr data field

MIKEYKVType
The key validity type

MIKEYKeyDataType
The type of key.

MIKEYMacAlg
Specifies the authentication algorithm used

MIKEYMapType
Specifies the method of uniquely mapping Crypto Sessions to the security protocol sessions.

MIKEYPRFFunc
The PRF function that has been/will be used for key derivation

MIKEYPayloadType
Different MIKEY Payload types.

MIKEYSecProto
Specifies the security protocol

MIKEYSecSRTP
This policy specifies the parameters for SRTP and SRTCP

MIKEYTSType
Specifies the timestamp type.

MIKEYType
Different MIKEY data types.

SDPResult
Return values for the SDP functions.

Type
A GLib or GLib-based library type

Traits

Cast
Upcasting and downcasting support.

IsA
Declares the "is a" relationship.

StaticType
Types that are supported by GLib dynamic typing.

ToValue
Converts to Value.