Struct gstreamer_sdp::MIKEYMessage

pub struct MIKEYMessage(_);

Structure holding the information of the MIKEY message

Methods

impl MIKEYMessage

pub fn new() -> MIKEYMessage

Make a new MIKEY message.

Returns

a new MIKEYMessage on success

pub fn new_from_caps(caps: &Caps) -> MIKEYMessage

Makes mikey message including:

Security Policy Payload

Key Data Transport Payload

Key Data Sub-Payload

caps

a gst::Caps, including SRTP parameters (srtp/srtcp cipher, authorization, key data)

Returns

a MIKEYMessage, or None if there is no srtp information in the caps.

pub fn add_cs_srtp(&mut self, policy: u8, ssrc: u32, roc: u32) -> bool

Add a Crypto policy for SRTP to self.

policy

The security policy applied for the stream with ssrc

ssrc

the SSRC that must be used for the stream

roc

current rollover counter

Returns

true on success

pub fn add_pke(&mut self, C: MIKEYCacheType, data: &[u8]) -> bool

Add a new PKE payload to self with the given parameters.

C

envelope key cache indicator

data_len

the length of data

data

the encrypted envelope key

Returns

true on success

pub fn add_rand(&mut self, rand: &[u8]) -> bool

Add a new RAND payload to self with the given parameters.

len

the length of rand

rand

random data

Returns

true on success

pub fn add_rand_len(&mut self, len: u8) -> bool

Add a new RAND payload to self with len random bytes.

len

length

Returns

true on success

pub fn add_t_now_ntp_utc(&mut self) -> bool

Add a new T payload to self that contains the current time in NTP-UTC format.

Returns

true on success

pub fn find_payload(
&self,
type_: MIKEYPayloadType,
nth: u32
) -> Option<MIKEYPayload>

Find the nth occurence of the payload with type_ in self.

type_

a MIKEYPayloadType

nth

payload to find

Returns

the nth MIKEYPayload of type_.

pub fn get_n_cs(&self) -> u32

Get the number of crypto sessions in self.

Returns

the number of crypto sessions

pub fn get_n_payloads(&self) -> u32

Get the number of payloads in self.

Returns

the number of payloads in self

pub fn remove_cs_srtp(&mut self, idx: i32) -> bool

Remove the SRTP policy at idx.

idx

the index to remove

Returns

true on success

pub fn remove_payload(&mut self, idx: u32) -> bool

Remove the payload in self at idx

idx

an index

Returns

true on success

pub fn set_info(
&mut self,
version: u8,
type_: MIKEYType,
V: bool,
prf_func: MIKEYPRFFunc,
CSB_id: u32,
map_type: MIKEYMapType
) -> bool

Set the information in self.

version

a version

type_

a MIKEYType

V

verify flag

prf_func

the MIKEYPRFFunc function to use

CSB_id

the Crypto Session Bundle id

map_type

the GstMIKEYCSIDMapType

Returns

true on success

pub fn to_caps(&self, caps: &Caps) -> bool

Feature: v1_8_1

caps

a gst::Caps to be filled with SRTP parameters (srtp/srtcp cipher, authorization, key data)

Returns

true on success